// src/app/models/sjcs/PretensionItem.ts
import { ComboItem } from '../../features/administracion/parametros/parametros-generales/parametros-generales.component';
export class PretensionItem {
  idPretension: string;
  descripcion: string;
  idInstitucion: string;
  fechaModificacion: string;
  fechabaja: string;
  codigoExt: string;
  idJurisdiccion: string;
  historico: boolean;
  descripcionJurisdiccion: string;
  constructor() { }
}